requirements.txt
python-pptx>=0.6.21
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.1 